+++ tasks/spec/expected/var/flows/my-flow/dependencies.js
@@ -3,6 +3,7 @@ define( [
    'laxar-path-controls/ax-test-control/ax-new-style-control',
    'laxar-application/widgets/default/local_widget/local_widget',
    'laxar-application/widgets/default/test_widget/test_widget',
+   'bower-widget/bower-widget',
    'laxar-application/widgets/default/plain_widget/plain_widget'
 ], function() {
    'use strict';
@@ -10,6 +11,6 @@ define( [
    var modules = [].slice.call( arguments );
    return {
       'angular': modules.slice( 0, 4 ),
-      'plain': modules.slice( 4, 5 )
+      'plain': modules.slice( 4, 6 )
    };
 } );